Division in modules
