+++ src/Emulator/Chippotto.js
@@ -1,24 +1,28 @@
 import Memory from "./Memory.js";
+import Screen from "./Screen.js";
+import Processor from "./CPU.js";
 
-class Chippotto{
-    constructor(Screen, LoadRom){
+class Chippotto {
+    constructor(screen, LoadRom) {
 
-        this.Screen = Screen
-       
-        this.Rom = LoadRom  
-        this.Rom.addEventListener("click", startRom())
-        
-        
-        this.Memory = Memory()
+        this.Memory = new Memory()
+        this.Screen = new Screen(screen)
+        this.Processor = new Processor(this.Memory, this.Screen)
 
+        this.Rom = LoadRom
+        this.initializeEmulator()
     }
-    startRom(){
-        fetch('./IgnoreForTest/chip8-roms-master/chip8-roms-master/programs/IBM Logo.ch8')
-        .then((data) =>data.blob())
-        .then((c) => c.arrayBuffer())
-        .then((aB) => {
-            code = new Uint8Array (aB)
-            console.log(code)
+
+    initializeEmulator() {
+        this.Rom.addEventListener("click", () => {
+            fetch('./../../IgnoreForTest/chip8-roms-master/chip8-roms-master/programs/IBM Logo.ch8')
+                .then((data) => data.blob())
+                .then((c) => c.arrayBuffer())
+                .then((aB) => {
+                    let code = new Uint8Array(aB)
+                    this.Memory.loadRom(code)
+                    this.Processor.run()
+                })
         })
     }
 }
