+++ src/Interface/index.html
@@ -12,5 +12,5 @@
         Load Rom
    </button>
 </body>
-<script src="script.js" defer></script>
+<script type="module" src="script.js" defer></script>
 </html>
